+++ net_util/Cargo.toml
@@ -5,11 +5,11 @@ authors = ["The Chromium OS Authors"]
 
 [dependencies]
 libc = "0.2.72"
+net_gen = { path = "../net_gen" }
 rand = "0.7.3"
 serde = "1.0.114"
+virtio-bindings = "0.1.0"
 vmm-sys-util = ">=0.3.1"
-
-net_gen = { path = "../net_gen" }
 
 [dev-dependencies]
 lazy_static = "1.3.0"
